Remove command line prompt from cli examples
GitHub automatically provides a copy for any code block meant to
be an example of a command to run from the command line.
We sometimes show the command and its output.  But for all the
cases where we just give the command to run, remove any
cli prompt character so that they can use the GitHub copy icon to
copy / paste the command.